requirements.txt
streamlit
speechrecognition
langchain
langchain-community
langchain-huggingface
langchain-groq
faiss-cpu
python-dotenv
sentence-transformers
arxiv
wikipedia
